fix job

+++ prn_job_manager/Models/QrtzSimpleTrigger.cs
@@ -8,9 +8,9 @@ namespace prn_job_manager.Models
         public string SchedName { get; set; } = null!;
         public string TriggerName { get; set; } = null!;
         public string TriggerGroup { get; set; } = null!;
-        public long RepeatCount { get; set; }
+        public int RepeatCount { get; set; }
         public long RepeatInterval { get; set; }
-        public long TimesTriggered { get; set; }
+        public int TimesTriggered { get; set; }
 
         public virtual QrtzTrigger QrtzTrigger { get; set; } = null!;
     }
